# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateScribeJobRequest": {
        "properties": {
          "idempotency_key": {
            "maxLength": 200,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "title": "Idempotency Key",
            "type": "string"
          },
          "interaction_type": {
            "enum": [
              "doctor_consult",
              "nurse_consult",
              "ai_patient_session"
            ],
            "title": "Interaction Type",
            "type": "string"
          },
          "source_record_id": {
            "format": "uuid",
            "title": "Source Record Id",
            "type": "string"
          }
        },
        "required": [
          "source_record_id",
          "interaction_type",
          "idempotency_key"
        ],
        "title": "CreateScribeJobRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportanceFeedbackRequest": {
        "properties": {
          "clinic_id": {
            "format": "uuid",
            "title": "Clinic Id",
            "type": "string"
          },
          "event_id": {
            "format": "uuid",
            "title": "Event Id",
            "type": "string"
          },
          "feedback_kind": {
            "enum": [
              "accept",
              "reject",
              "pin",
              "edit",
              "comment"
            ],
            "title": "Feedback Kind",
            "type": "string"
          },
          "topic": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "clinic_id",
          "topic",
          "feedback_kind"
        ],
        "title": "ImportanceFeedbackRequest",
        "type": "object"
      },
      "ImportancePreferenceResponse": {
        "properties": {
          "clinic_id": {
            "format": "uuid",
            "title": "Clinic Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "profile_id": {
            "format": "uuid",
            "title": "Profile Id",
            "type": "string"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "weight": {
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "id",
          "clinic_id",
          "profile_id",
          "topic",
          "weight",
          "updated_at"
        ],
        "title": "ImportancePreferenceResponse",
        "type": "object"
      },
...
      "ProviderUsageResponse": {
        "properties": {
          "average_latency_ms": {
            "title": "Average Latency Ms",
            "type": "number"
          },
          "calls": {
            "title": "Calls",
            "type": "integer"
          },
          "estimated_cost_usd": {
            "title": "Estimated Cost Usd",
            "type": "number"
          },
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "model",
          "calls",
          "input_tokens",
          "output_tokens",
          "average_latency_ms",
          "estimated_cost_usd"
        ],
        "title": "ProviderUsageResponse",
        "type": "object"
      },
...
      "ScribeJobEventResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "event_kind": {
            "enum": [
              "generating",
              "validating",
              "persisting",
              "completed",
              "retrying",
              "cancelled"
            ],
            "title": "Event Kind",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "job_id",
          "event_kind",
          "created_at"
        ],
        "title": "ScribeJobEventResponse",
        "type": "object"
      },
      "ScribeJobResponse": {
        "properties": {
          "attempt_count": {
            "title": "Attempt Count",
            "type": "integer"
          },
          "available_at": {
            "format": "date-time",
            "title": "Available At",
            "type": "string"
          },
          "claimed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimed At"
          },
          "clinic_id": {
            "format": "uuid",
            "title": "Clinic Id",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "estimated_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost Usd"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "interaction_type": {
            "enum": [
              "doctor_consult",
              "nurse_consult",
              "ai_patient_session"
            ],
            "title": "Interaction Type",
            "type": "string"
          },
          "lease_expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lease Expires At"
          },
          "max_attempts": {
            "title": "Max Attempts",
            "type": "integer"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "output_entry_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Entry Id"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "patient_id": {
            "format": "uuid",
            "title": "Patient Id",
            "type": "string"
          },
          "provider_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Name"
          },
          "queue_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Position"
          },
          "requested_by": {
            "format": "uuid",
            "title": "Requested By",
            "type": "string"
          },
          "safe_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safe Error Code"
          },
          "source_record_id": {
            "format": "uuid",
            "title": "Source Record Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "processing",
              "succeeded",
              "failed",
              "dead_letter",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "clinic_id",
          "patient_id",
          "source_record_id",
          "interaction_type",
          "requested_by",
          "idempotency_key",
          "status",
          "attempt_count",
          "max_attempts",
          "available_at",
          "claimed_at",
          "lease_expires_at",
          "completed_at",
          "safe_error_code",
          "output_entry_id",
          "created_at",
          "updated_at"
        ],
        "title": "ScribeJobResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Server-enforced API for the Nightingale longitudinal Care Note.",
    "title": "Nightingale Care Note API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/importance-feedback": {
      "post": {
        "operationId": "record_importance_feedback_importance_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportanceFeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportancePreferenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record Importance Feedback",
        "tags": [
          "adaptive importance"
        ]
      }
    },
    "/importance-preferences": {
      "get": {
        "operationId": "list_importance_preferences_importance_preferences_get",
        "parameters": [
          {
            "in": "query",
            "name": "clinic_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Clinic Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ImportancePreferenceResponse"
                  },
                  "title": "Response List Importance Preferences Importance Preferences Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Importance Preferences",
        "tags": [
          "adaptive importance"
        ]
      }
    },
    "/importance-preferences/{clinic_id}": {
      "delete": {
        "operationId": "reset_importance_preferences_importance_preferences__clinic_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "clinic_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Clinic Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Reset Importance Preferences Importance Preferences  Clinic Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reset Importance Preferences",
        "tags": [
          "adaptive importance"
        ]
      }
    },
...
    "/patients/{patient_id}/scribe-job-events": {
      "get": {
        "operationId": "list_patient_scribe_job_events_patients__patient_id__scribe_job_events_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScribeJobEventResponse"
                  },
                  "title": "Response List Patient Scribe Job Events Patients  Patient Id  Scribe Job Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Patient Scribe Job Events",
        "tags": [
          "AI scribe jobs"
        ]
      }
    },
    "/patients/{patient_id}/scribe-jobs": {
      "get": {
        "operationId": "list_patient_scribe_jobs_patients__patient_id__scribe_jobs_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScribeJobResponse"
                  },
                  "title": "Response List Patient Scribe Jobs Patients  Patient Id  Scribe Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Patient Scribe Jobs",
        "tags": [
          "AI scribe jobs"
        ]
      },
      "post": {
        "operationId": "create_scribe_job_patients__patient_id__scribe_jobs_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScribeJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScribeJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Scribe Job",
        "tags": [
          "AI scribe jobs"
        ]
      }
    },
...
    "/provider-usage": {
      "get": {
        "operationId": "provider_usage_dashboard_provider_usage_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProviderUsageResponse"
                  },
                  "title": "Response Provider Usage Dashboard Provider Usage Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Provider Usage Dashboard",
        "tags": [
          "AI scribe jobs"
        ]
      }
    },
    "/scribe-jobs/{job_id}": {
      "get": {
        "operationId": "get_scribe_job_scribe_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScribeJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Scribe Job",
        "tags": [
          "AI scribe jobs"
        ]
      }
    },
    "/scribe-jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_scribe_job_scribe_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScribeJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Scribe Job",
        "tags": [
          "AI scribe jobs"
        ]
      }
    },
    "/scribe-jobs/{job_id}/queue-position": {
      "get": {
        "operationId": "get_scribe_job_queue_position_scribe_jobs__job_id__queue_position_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "title": "Response Get Scribe Job Queue Position Scribe Jobs  Job Id  Queue Position Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Scribe Job Queue Position",
        "tags": [
          "AI scribe jobs"
        ]
      }
    },
...
  }
}